requirements.txt
# Python 3.10.11

openai==1.13.3
streamlit==1.32.0
audio_recorder_streamlit
streamlit-float
python-dotenv
langchain==0.1.16
chromadb
tiktoken
pypdf==4.2.0
yt_dlp
pydub==0.25.1
librosa==0.10.1